### 2021-09-16 #80 @whedon
[PRE REVIEW]: Catwalk.jl: An adaptive dispatch optimizer
**Submitting author:** @tisztamo (<a href="http://orcid.org/0000-0003-4931-4791">Krisztián Schäffer</a>)
**Repository:** <a href="https://github.com/tisztamo/Catwalk.jl" target ="_blank">https://github.com/tisztamo/Catwalk.jl</a>
**Version:**
