### 2021-09-16 #80 @tisztamo
The paper is in a separate paper branch, as described in the guide: https://juliacon.github.io/proceedings-guide/author/#submission_details

Should I move it to the main branch?
