+++ paper/paper.tex
@@ -11,11 +11,11 @@
 
 \begin{abstract}
 
-\verb Catwalk.jl  is a JIT compiler implemented as a Julia library that generates optimized dispatch code based on statistical profiling.
+\verb|Catwalk.jl| is a JIT compiler implemented as a Julia library that generates optimized dispatch code based on statistical profiling.
 Unlike typical JIT compilers it requires some integration work from its users, allowing it to completely eliminate
 the need of complex deoptimization logic. It is able to compile new type-stabilized routes or reorder existing ones if the distribution of dispatched types changes during runtime and the customizable cost model predicts significant speedup compared to the best version that was previously compiled.
 
-\verb Catwalk.jl  was designed for situations when both composability and run-time polymorphism is required,
+\verb|Catwalk.jl| was designed for situations when both composability and run-time polymorphism is required,
 and some run-time compilation overhead is acceptable for speeding up dynamic dispatch in hot loops.
 
 \headingtable
@@ -30,7 +30,7 @@ A variety of techniques and packages is known to alleviate this issue, but they 
 dependent packages cannot benefit from the speedup \cite{unionsplitting} \cite{manualdispatch}.
 
 
-\verb Catwalk.jl  uses a technique I call "iterated staging" to include all encountered types in the optimization, while interacting with the Julia compiler using only standard metaprogramming facilities: macros, \verb @generated  functions and generics.
+\verb|Catwalk.jl| uses a technique I call "iterated staging" to include all encountered types in the optimization, while interacting with the Julia compiler using only standard metaprogramming facilities: macros, \verb|@generated| functions and generics.
 
 \section{Integration: Iterated staging}
 
@@ -46,14 +46,14 @@ for batch_num in 1:BATCH_COUNT
 end
 \end{lstlisting}
 
-\verb single_batch  or its downstream function that contains the dynamically dispatched call site must also be marked with the \verb @jit  macro that turns
+\verb|single_batch|  or its downstream function that contains the dynamically dispatched call site must also be marked with the \verb|@jit|  macro that turns
 it into a generated function.
 
 \section {Compilation}
 
-This structure allows \verb Catwalk.jl  to recompile between batches if needed.
-The \emph{type} of the "JIT context" returned by the \verb Catwalk.ctx()  call
-determines the code to be generated. The call \verb f(a, unstable, b)  will 
+This structure allows \verb|Catwalk.jl| to recompile between batches if needed.
+The \emph{type} of the "JIT context" returned by the \verb|Catwalk.ctx()| call
+determines the code to be generated. The call \verb|f(a, unstable, b)| will 
 be rewritten to:
 
 \begin{lstlisting}[language = Julia]
@@ -79,11 +79,11 @@ of stabilized types, the cost model, the profiler and the optimizer.
 
 \section {Conclusion and further work}
 
-Thanks to the "Just Ahead Of Time" compilation model of Julia, \verb Catwalk.jl  can adaptively generate highly optimal dispatch code, while working completely in "user space".
+Thanks to the "Just Ahead Of Time" compilation model of Julia, \verb|Catwalk.jl| can adaptively generate highly optimal dispatch code, while working completely in "user space".
 
 Compilation cost is effectively minimized by manual impact point selection and backtesting peviously compiled versions against newly collected profiles using a cost model of dispatch. 
 
-\verb Catwalk.jl  speeds up dynamic dispatch by a factor typically between 5 and 200. Total experienced speedup exceeds 30\% in some real-life programs.
+\verb|Catwalk.jl| speeds up dynamic dispatch by a factor typically between 5 and 200. Total experienced speedup exceeds 30\% in some real-life programs.
 
 In the future I plan to improve compilation overhead by using more efficient type-encoding;
[…]
