### 2021-09-17 #80 @tisztamo
Thank you, @carstenbauer !

I have fixed some overleaf-related incompatibility and checked locally using a fresh MacTex. It seem to work. Not sure, how I can continue. E.g. Where can I find the logs?

---
My logs:

ko@MacBook-Pro-3 tmp % git clone https://github.com/tisztamo/Catwalk.jl.git
Cloning into 'Catwalk.jl'...
remote: Enumerating objects: 791, done.
remote: Total 791 (delta 406), reused 615 (delta 252), pack-reused 0
Resolving deltas: 100% (406/406), done.
ko@MacBook-Pro-3 tmp % cd Catwalk.jl
ko@MacBook-Pro-3 Catwalk.jl % git checkout paper
Branch paper set up to track remote branch paper from origin.
Switched to a new branch 'paper'
ko@MacBook-Pro-3 Catwalk.jl % cd paper
ko@MacBook-Pro-3 paper % pdflatex  -recorder  "paper.tex"
This is pdfTeX, Version 3.141592653-2.6-1.40.22 (TeX Live 2021) (preloaded format=pdflatex)
 restricted \write18 enabled.
entering extended mode
(./paper.tex
LaTeX2e <2020-10-01> patch level 4
L3 programming layer <2021-02-18> (./juliacon.cls
(/usr/local/texlive/2021/texmf-dist/tex/latex/psnfss/helvet.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/keyval.sty))
Document Class: juliacon
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/latexsym.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/url/url.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/inputenc.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/fontenc.sty)
(./journal_dat.tex)
(/usr/local/texlive/2021/texmf-dist/tex/latex/psnfss/times.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/tools/bm.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/graphicx.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/graphics.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/trig.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-cfg/graphics.cfg)
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-def/pdftex.def)))
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/hyperref.sty
(/usr/local/texlive/2021/texmf-dist/tex/generic/ltxcmds/ltxcmds.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/iftex/iftex.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/pdftexcmds/pdftexcmds.sty
(/usr/local/texlive/2021/texmf-dist/tex/generic/infwarerr/infwarerr.sty))
(/usr/local/texlive/2021/texmf-dist/tex/generic/kvsetkeys/kvsetkeys.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/kvdefinekeys/kvdefinekeys.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/pdfescape/pdfescape.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/hycolor/hycolor.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/letltxmacro/letltxmacro.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/auxhook/auxhook.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/kvoptions/kvoptions.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/pd1enc.def)
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/hyperref-langpatches.def
) (/usr/local/texlive/2021/texmf-dist/tex/generic/intcalc/intcalc.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/etexcmds/etexcmds.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/puenc.def)
(/usr/local/texlive/2021/texmf-dist/tex/generic/bitset/bitset.sty
(/usr/local/texlive/2021/texmf-dist/tex/generic/bigintcalc/bigintcalc.sty))
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/atbegshi-ltx.sty))
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/hpdftex.def
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/atveryend-ltx.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/rerunfilecheck/rerunfilecheck.sty

(/usr/local/texlive/2021/texmf-dist/tex/generic/uniquecounter/uniquecounter.sty
))) (./jlcode.sty
-- Package: `jlcode' 2.1 <2018/03/06> --
(/usr/local/texlive/2021/texmf-dist/tex/latex/listings/listings.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/listings/lstmisc.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/listings/listings.cfg))
(/usr/local/texlive/2021/texmf-dist/tex/latex/xcolor/xcolor.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-cfg/color.cfg))
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/textcomp.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/amssymb.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/amsfonts.sty))
(/usr/local/texlive/2021/texmf-dist/tex/latex/eurosym/eurosym.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/fontenc.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/psnfss/t1ptm.fd))
(/usr/local/texlive/2021/texmf-dist/tex/latex/tools/calc.sty)
(/usr/local/texlive/2021/texmf-dist/tex/latex/base/t1cmtt.fd))
(/usr/local/texlive/2021/texmf-dist/tex/latex/preprint/authblk.sty))
(/usr/local/texlive/2021/texmf-dist/tex/latex/l3backend/l3backend-pdftex.def)
No file paper.aux.
(/usr/local/texlive/2021/texmf-dist/tex/context/base/mkii/supp-pdf.mkii
[Loading MPS to PDF converter (version 2006.09.02).]
) (/usr/local/texlive/2021/texmf-dist/tex/latex/epstopdf-pkg/epstopdf-base.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/latexconfig/epstopdf-sys.cfg))
(/usr/local/texlive/2021/texmf-dist/tex/latex/hyperref/nameref.sty
(/usr/local/texlive/2021/texmf-dist/tex/latex/refcount/refcount.sty)
(/usr/local/texlive/2021/texmf-dist/tex/generic/gettitlestring/gettitlestring.s
ty)) (./header.tex) (/usr/local/texlive/2021/texmf-dist/tex/latex/base/ulasy.fd
) (/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/umsa.fd)
(/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/umsb.fd)

LaTeX Warning: Citation `subtypeperf' on page 1 undefined on input line 27.

LaTeX Warning: Citation `typesortedcollections' on page 1 undefined on input li
ne 29.

LaTeX Warning: Citation `singledispatcharrays' on page 1 undefined on input lin
e 29.

LaTeX Warning: Citation `unionsplitting' on page 1 undefined on input line 30.

LaTeX Warning: Citation `manualdispatch' on page 1 undefined on input line 30.

Underfull \hbox (badness 2818) in paragraph at lines 54--58
\T1/ptm/m/n/9 if needed. The \T1/ptm/m/it/9 type \T1/ptm/m/n/9 of the "JIT con-
text" re-turned by the
(./bib.tex (./paper.bbl
Underfull \hbox (badness 10000) has occurred while \output is active
  []

Overfull \vbox (2.39996pt too high) has occurred while \output is active
[1{/usr/local/texlive/2021/texmf-var/fonts/map/pdftex/updmap/pdftex.map}
Non-PDF special ignored!
<special> papersize=8.5in,11in
 <./logojuliacon.pdf>]
Underfull \hbox (badness 10000) in paragraph at lines 9--12
[]\T1/ptm/m/n/9 jlapeyre@github John Lapeyre. Sin-gledis-patchar-rays.jl.

Underfull \hbox (badness 10000) in paragraph at lines 24--27
[]\T1/ptm/m/n/9 tkoolen@github Twan Koolen. Type-sort-ed-col-lec-tions.jl.
))
Overfull \vbox (2.39996pt too high) has occurred while \output is active
Underfull \hbox (badness 10000) has occurred while \output is active
  \T1/ptm/m/n/9 2
[2] (./paper.aux)

LaTeX Warning: There were undefined references.

LaTeX Warning: Label(s) may have changed. Rerun to get cross-references right.

Package rerunfilecheck Warning: File `paper.out' has changed.
(rerunfilecheck)                Rerun to get outlines right
(rerunfilecheck)                or use package `bookmark'.

 )
(see the transcript file for additional information){/usr/local/texlive/2021/te
xmf-dist/fonts/enc/dvips/cm-super/cm-super-t1.enc}{/usr/local/texlive/2021/texm
f-dist/fonts/enc/dvips/base/8r.enc}</usr/local/texlive/2021/texmf-dist/fonts/ty
pe1/public/amsfonts/latxfont/lcircle1.pfb></usr/local/texlive/2021/texmf-dist/f
onts/type1/public/cm-super/sftt0800.pfb></usr/local/texlive/2021/texmf-dist/fon
ts/type1/public/cm-super/sftt0900.pfb></usr/local/texlive/2021/texmf-dist/fonts
/type1/public/cm-super/sftt1000.pfb></usr/local/texlive/2021/texmf-dist/fonts/t
ype1/urw/times/utmb8a.pfb></usr/local/texlive/2021/texmf-dist/fonts/type1/urw/t
imes/utmr8a.pfb></usr/local/texlive/2021/texmf-dist/fonts/type1/urw/times/utmri
8a.pfb>
Output written on paper.pdf (2 pages, 136836 bytes).
Transcript written on paper.log.
