+++ paper/paper.yml
@@ -1,9 +1,10 @@
 title: "My JuliaCon proceeding"
 keywords:
   - Julia
-  - Optimization
-  - Game theory
-  - Compiler
+  - JIT
+  - Dynamic Dispatch
+  - Adaptive Optimization
+  - Profiler
 authors:
   - name: Krisztián Schäffer
     orcid: 0000-0003-4931-4791
