### 2021-10-07 #80 @ranjanan
@tisztamo update https://github.com/tisztamo/Catwalk.jl/blob/paper/paper/paper.yml please.
