+++ paper/paper.yml
@@ -1,4 +1,4 @@
-title: "My JuliaCon proceeding"
+title: "Catwalk.jl: An adaptive dispatch optimizer"
 keywords:
   - Julia
   - JIT
