### 2021-12-02 #98 @whedon
[REVIEW]: Catwalk.jl: An adaptive dispatch optimizer
**Submitting author:** @tisztamo (<a href="http://orcid.org/0000-0003-4931-4791">Krisztián Schäffer</a>)
**Repository:** <a href="https://github.com/tisztamo/Catwalk.jl" target ="_blank">https://github.com/tisztamo/Catwalk.jl</a>
**Version:** paper_juliacon_21

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
