### 2021-12-02 #98 @Roger-luo
This package looks nice in general to me (I actually tried it when it was first announced in discourse), the interface is simple to use and the performance gain is fantastic. The only thing that is missing is the community guidelines required above, I'd suggest just putting a https://github.com/SciML/ColPrac badge in the README
