+++ README.md
@@ -8,6 +8,7 @@
 ![Lifecycle](https://img.shields.io/badge/lifecycle-dormant-blue.svg) -->
 [![Documentation](https://img.shields.io/badge/docs-dev-blue.svg)](https://tisztamo.github.io/Catwalk.jl/dev/)
 [![CI](https://github.com/tisztamo/Catwalk.jl/actions/workflows/ci.yml/badge.svg)](https://github.com/tisztamo/Catwalk.jl/actions/workflows/ci.yml)
+[![ColPrac: Contributor Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac)
 
 Catwalk.jl can speed up long-running Julia processes by minimizing the
 overhead of dynamic dispatch. It is a JIT compiler that continuosly
