### 2021-12-03 #98 @tisztamo
Thank you very much @Roger-luo for the review!

fix of the missing guideline: https://github.com/tisztamo/Catwalk.jl/commit/300518aaa0fed31d7a5f25a2fb5de220eec8a11c
