### 2021-12-04 #98 @aviatesk
I read the article, and here are few feedbacks on details:
- For the sentence "Types injected from dependent packages cannot benefit from the speedup" in `1. Introduction`: Union-splitting is applied for general Julia programs, so it optimizes arbitrary code. So I'd say the reference of `[1]` isn't much appropriate there
  - also, I couldn't find a reachable reference of `[2] Singledispatcharrays.jl`/`[5] Singledispatcharrays.jl` anymore
- maybe it's more appropriated if `2. Integration: Iterated staging` has some more detailed explanation of the usage (like as explained at <https://tisztamo.github.io/Catwalk.jl/dev/usage/>)?

---

As a general feedback, it might be more interesting if we can explain the essential difference between union-splitting (static optimization of runtime dispatches) vs. Catwalk's profiling based optimization.
The biggest difference would be that the former can only optimize it when all possible types are known AOT, while the latter can interact with runtime information. This means, in some situations, union-splitting is just enough. For example, if we change the definition of `xs` in [Catwalk's documentation](https://tisztamo.github.io/Catwalk.jl/dev/usage/) to `const xs = Union{Int,Float64,ComplexF64}[1, 2.0, ComplexF64(3.0, 3.0)]`, there is no essential performance difference:
```julia
julia> const NUM_BATCHES = 1000
1000

julia> function runbatches()
           for batchidx = 1:NUM_BATCHES
               hotloop()
               # Log progress, etc.
           end
       end
runbatches (generic function with 1 method)

julia>

julia> const NUM_ITERS_PER_BATCH = 1_000_000
1000000

julia> function hotloop()
           for i = 1:NUM_ITERS_PER_BATCH
               x = get_some_x(i)
               calc_with_x(x)
           end
       end
hotloop (generic function with 1 method)

julia> const xs = Union{Int,Float64,ComplexF64}[1, 2.0, ComplexF64(3.0, 3.0)]
3-element Vector{Union{Float64, Int64, ComplexF64}}:
     1
     2.0
 3.0 + 3.0im

julia> get_some_x(i) = xs[i % length(xs) + 1]
get_some_x (generic function with 1 method)

julia> const result = Ref(ComplexF64(0.0, 0.0))
Base.RefValue{ComplexF64}(0.0 + 0.0im)

julia> function calc_with_x(x)
           result[] += x
       end
calc_with_x (generic function with 1 method)

julia> using Catwalk

julia> @jit calc_with_x x function hotloop_jit(jitctx)
           for i = 1:NUM_ITERS_PER_BATCH
               x = get_some_x(i)
               calc_with_x(x)
           end
       end
hotloop_jit (generic function with 1 method)

julia> function runbatches_jit()
           jit = Catwalk.JIT() ## Also works inside a function (no eval used)
           for batch = 1:NUM_BATCHES
               Catwalk.step!(jit)
               hotloop_jit(Catwalk.ctx(jit)) # Dynamic dispatch is "pulled out" from the hot loop to this call
           end
       end
runbatches_jit (generic function with 1 method)

julia> result[] = ComplexF64(0, 0)
0.0 + 0.0im

julia> @time runbatches_jit()
  4.193906 seconds (2.63 M allocations: 145.252 MiB, 0.50% gc time, 25.96% compilation time)

julia> jit_result = result[]
2.0e9 + 9.99999e8im

julia> result[] = ComplexF64(0, 0)
0.0 + 0.0im

julia> @time runbatches()
  2.854675 seconds
```
And then we may also want to mention the heuristic of Julia compiler, e.g. we should use Catwalk when we have more than [5 possible types](https://github.com/JuliaLang/julia/blob/83d7ebaca528904e7bcfc7f032f19d44fc1acd49/base/compiler/types.jl#L122) :)
