### 2021-12-04 #98 @aviatesk
If you like a more complicated story, we can also mention a part of Julia compiler where it does some sort of statical profiling JIT. It actually sorts method table by the runtime counts of dispatch, so Julia runtime actually does something very similar to Catwalk's idea (well, even in that case `isa`-based code generated by Catwalk is far more efficient than the whole method dispatch). So we can also make an advanced discussion on the computation costs of `isa` checks and runtime dispatch invocation, where the latter is around 20 CPU-cycles when dealing with a single argument, while the former is around 1-2 CPU cycles when consulted with a concrete type, if I remember correctly.
