### 2021-12-05 #98 @tisztamo
Thank you very much for your valuable comments, @aviatesk !

> * For the sentence "Types injected from dependent packages cannot benefit from the speedup" in `1. Introduction`: Union-splitting is applied for general Julia programs, so it optimizes arbitrary code. So I'd say the reference of `[1]` isn't much appropriate there

Right, this is not true for union splitting, I have to correct it (true for ManualDispatch.jl).

Still the inability to extend unions is prohibitive in some cases. Your later wording "can only optimize it when all possible types are known AOT" better describes the difference, but the precise meaning of AOT is not exactly clear to me. If we assume that a package may implement iterated staging combined with union split-based code generation - which it can do in a few dozen lines, although no package did it before Catwalk - , then the only differences are IIUC the limit of 5, and that Union{A, B, Any} will not be split. Do you mean this latter limitation with "AOT"?

> * also, I couldn't find a reachable reference of `[2] Singledispatcharrays.jl`/`[5] Singledispatcharrays.jl` anymore

It seems that the url field of references is not rendered, I will fix it.

https://github.com/tkoolen/TypeSortedCollections.jl
https://github.com/melonedo/SingleDispatchArrays.jl

> in some situations, union-splitting is just enough

Yes, that is a contrived example, I will add a note to the docs. (The [Alternatives](https://tisztamo.github.io/Catwalk.jl/dev/#Alternatives) section of the introduction mentions union splitting among other solutions.)

> And then we may also want to mention the heuristic of Julia compiler, e.g. we should use Catwalk when we have more than [5 possible types](https://github.com/JuliaLang/julia/blob/83d7ebaca528904e7bcfc7f032f19d44fc1acd49/base/compiler/types.jl#L122) :)

I will find space for a sentence about it. (The one-page limitation makes it very hard to speak about other solutions, I thought that referencing them is enough.)

> Julia runtime actually does something very similar to Catwalk's idea

I did not know of the method table sorting, thanks for mentioning it! As I am not allowed to run out of the page - or I need to write 4+ more which I cannot do now -, unfortunately I have to leave it out for now.

> So we can also make an advanced discussion on the computation costs of isa checks and runtime dispatch invocation

The default cost model of Catwalk ( https://tisztamo.github.io/Catwalk.jl/dev/tuning/#Customize-the-cost-model ) is ad-hoc, it would be great to improve it! Discussing the relevant parts of the Julia compiler in a paper seems also helpful for the community as there is hard to find information on the internals. I hope to find time for a next iteration of Catwalk during the next year and I will contact you before starting if you don't mind.

Do you find the here outlined changes satisfactory?

Thank you for your help!
