### 2021-12-05 #98 @aviatesk
Yep, I think what you outlined above sounds good as a whole. I actually didn't think about the page limitation at all, so please summarize as you like -- my feedbacks are just suggestions :)

> Do you mean this latter limitation with "AOT"?

By "AOT", I simply meant when Julia compiler compiles a hot function in question.
There are few cases when it fails to union-split (i.e. optimize abstract callsite):
- when the type in question is an "abstract" type, e.g. `Any`, `Number`, `Union{Number,String}`
- when the type is union and it needs to be more than 5 concrete types when split

In these cases, I think Catwalk can optimize them further by iterated staging using runtime information.

> Discussing the relevant parts of the Julia compiler in a paper seems also helpful for the community as there is hard to find information on the internals. I hope to find time for a next iteration of Catwalk during the next year and I will contact you before starting if you don't mind.

Yep, that would be great. Looking forward to the collaboration :)
