+++ paper/header.tex
@@ -5,5 +5,5 @@
 \author[1]{Krisztián Schäffer}
 \affil[1]{independent researcher}
 
-\keywords{Julia, JIT, Profiler, Compiler, Adaptive Optimization}
+\keywords{Julia, JIT, Dynamic Dispatch, Adaptive Optimization, Profiler}
 
