### 2021-12-19 599e7b2
Add bib urls to note fields
