+++ paper/ref.bib
@@ -2,6 +2,7 @@
   title={Performance drawback with subtyping},
   author={Ronneesley Moura Teles},
   url={https://discourse.julialang.org/t/performance-drawback-with-subtyping/51939},
+  note={https://discourse.julialang.org/t/performance-drawback-with-subtyping/51939},
   journal={discourse.julialang.org},
   year={2020},
 }
@@ -10,6 +11,7 @@
   title={Union-splitting: what it is, and why you should care},
   author={Tim Holy},
   url={https://julialang.org/blog/2018/08/union-splitting/},
+  note={https://julialang.org/blog/2018/08/union-splitting/},
   journal={The Julia Language Blog},
   year={2018},
 }
@@ -18,6 +20,7 @@
   title={TypeSortedCollections.jl},
   author={Twan Koolen, tkoolen@github},
   url={https://github.com/tkoolen/TypeSortedCollections.jl},
+  note={https://github.com/tkoolen/TypeSortedCollections.jl},
   year={2017},
   journal={Github},
 }
@@ -26,6 +29,7 @@
   title={SingleDispatchArrays.jl},
   author={melonedo@github},
   url={https://github.com/melonedo/SingleDispatchArrays.jl},
+  note={https://github.com/melonedo/SingleDispatchArrays.jl},
   year={2021},
   journal={Github},
 }
@@ -34,6 +38,16 @@
   title={SingleDispatchArrays.jl},
   author={John Lapeyre, jlapeyre@github},
   url={https://github.com/jlapeyre/ManualDispatch.jl},
+  note={https://github.com/jlapeyre/ManualDispatch.jl},
+  year={2020},
+  journal={Github},
+}
+
+@article{catwalkwebsite,
+  title={Catwalk.jl},
+  author={Krisztián Schäffer, tisztamo@github},
+  url={https://github.com/tisztamo/Catwalk.jl},
+  note={https://github.com/tisztamo/Catwalk.jl},
   year={2020},
   journal={Github},
 }
