### 2021-12-19 106b91d
Fix false statement about union splitting + add link to the repo
