+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"usage/","page":"Usage","title":"Usage","text":"EditURL = \"https://github.com/tisztamo/Catwalk.jl/blob/main/docs/src/usage.jl\"","category":"page"},{"location":"usage/#Usage","page":"Usage","title":"Usage","text":"","category":"section"},{"location":"usage/","page":"Usage","title":"Usage","text":"Let's say you have a long-running calculation, organized into batches:","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"const NUM_BATCHES = 1000\n\nfunction runbatches()\n    for batchidx = 1:NUM_BATCHES\n        hotloop()\n        # Log progress, etc.\n    end\nend","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"The hot loop calls the type-unstable function get_some_x() and passes its result to a relatively cheap calculation calc_with_x().","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"const NUM_ITERS_PER_BATCH = 1_000_000\n\nfunction hotloop()\n    for i = 1:NUM_ITERS_PER_BATCH\n        x = get_some_x(i)\n        calc_with_x(x)\n    end\nend\n\nconst xs = Any[1, 2.0, ComplexF64(3.0, 3.0)]\nget_some_x(i) = xs[i % length(xs) + 1]\n\nconst result = Ref(ComplexF64(0.0, 0.0))\n\nfunction calc_with_x(x)\n    result[] += x\nend","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"As get_some_x is not type-stable, calc_with_x must be dynamically dispatched, which slows down the calculation.","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"Sometimes it is not feasible to type-stabilize get_some_x. Catwalk.jl is here for those cases.","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"You mark hotloop, the outer function with the @jit macro and provide the name of the dynamically dispatched function and the argument to operate on (the API will hopefully improve in the future). You also have to add an extra argument named jitctx to the jit-ed function:","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"using Catwalk\n\n@jit calc_with_x x function hotloop_jit(jitctx)\n    for i = 1:NUM_ITERS_PER_BATCH\n        x = get_some_x(i)\n        calc_with_x(x)\n    end\nend","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"The Catwalk optimizer will provide you the jitctx context which you have to pass to the jit-ed function manually. Also, every batch needs a bit housekeeping to drive the Catwalk optimizer:","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"function runbatches_jit()\n    jit = Catwalk.JIT() ## Also works inside a function (no eval used)\n    for batch = 1:NUM_BATCHES\n        Catwalk.step!(jit)\n        hotloop_jit(Catwalk.ctx(jit)) # Dynamic dispatch is \"pulled out\" from the hot loop to this call\n    end\nend","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"Note that the body of the batch can be arbitrarily complex, not just a simple for loop like in this example. See a more complex test case.","category":"page"},{"location":"usage/","page":"Usage","title":"Usage","text":"Yes, it is a bit complicated to integrate your code with Catwalk, but it may worth the effort:","category":"page"},{"location":"usage/","page
[…]
