### 2021-12-19 #98 @tisztamo
I made the required changes to the text:

https://github.com/tisztamo/Catwalk.jl/commit/106b91d22f42a598355055a0a159ab4e92b5da1f#diff-983b58b8186b0a4ed7f280f258cdab3eb0dd7d5136f8ac361ba982a43cfb7136

and updated the refs so that urls are rendered:

https://github.com/tisztamo/Catwalk.jl/commit/599e7b232e93c929aa6d1a1b4bf46d87bf8caffd#diff-382635c9325bf3273d195ff1b8a44e5b11afd7d97addeb8863ea35feb98c1a07

There is also a minimal change to the keywords:

https://github.com/tisztamo/Catwalk.jl/commit/e93afe2ef95143225f2cb7bf2b873f8b10b97071#diff-382635c9325bf3273d195ff1b8a44e5b11afd7d97addeb8863ea35feb98c1a07

@aviatesk Sorry for the delay again and thank you for reviewing the changes!
