### 2022-02-27 #98 @tisztamo
@ranjanan Could you please help me where to edit?

```
paper % grep 2019 *
journal_dat.tex:\def\@year{2019}
juliacon.cls:\def\filedate{2019 04 07}
Binary file logojuliacon.pdf matches
paper.log:Package: bm 2019/07/24 v1.2d Bold Symbol Support (DPC/FMi)
...
paper.log:Package: gettitlestring 2019/12/15 v1.6 Cleanup title references (HO)
prep.rb:year = ENV["JLCON_YEAR"] === nil ? 2019 : ENV["JLCON_YEAR"]
```

`journal_dat.tex` is generated. Should I edit the default in prep.rb? (It says "DO NOT EDIT").
