+++ paper/header.tex
@@ -7,3 +7,10 @@
 
 \keywords{Julia, JIT, Dynamic Dispatch, Adaptive Optimization, Profiler}
 
+\hypersetup{
+pdftitle = {Catwalk.jl: An adaptive dispatch optimizer},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {Krisztián Schäffer},
+pdfkeywords = {Julia, JIT, Dynamic Dispatch, Adaptive Optimization, Profiler},
+}
+
